feat(core): implement pipeline for upsert docs with binding file_id and user_id
